Updated Course.cs, Instructor.cs, Program.cs, Room.cs, Schedule.cs: Created ISchedule.cs
Updated each of the components to further build the base of the program before adding in the fitness function.

ISchedule.cs is the interface for Schedule.cs, and will allow for the design and implementation of the Fitness Function within the Schedule class.

Each class will write a set of lines to console when an object is created an initalized to allow for troubleshooting.

Currently I've been overloading two extra constructors, since my original overloaded constructor hasn't been easy to implement with the Schedule class receiving string input. However, I'm likely to remove the expansive overloaded constructors in favor of the smaller single string parameter constructors.

Currently, only the long form constructors are printing to the console when the objects are created, but hopefully I'll be able to complete that for the other ones soon too.